% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-preference-data.r
\name{fit_preference}
\alias{fit_preference}
\title{Fit the preference data collected from a clinical trial}
\usage{
fit_preference(outcome, random, treatment, strata)
}
\arguments{
\item{outcome}{(numeric) individual trial outcomes.}

\item{random}{(logical) was this individual part of the random arm?}

\item{treatment}{(character, factor, or integer) which treatment an 
individual received}

\item{strata}{(optional integer) which strata the individual belongs to.}
}
\description{
TODO fill this in.
}
\examples{
##Unstratified
outcome <- c(10, 8, 6, 10, 5, 8, 7, 6, 10, 12, 11, 6, 8, 10, 5, 7, 9, 12, 6,
             8, 9, 10, 7, 8,11)
random <- c(rep(FALSE, 13), rep(TRUE, 12))
treatment <- c(rep(1, 5), rep(2, 8), rep(1, 6), rep(2, 6))
fit_preference(outcome, random, treatment)

##Stratified
#x1 <- c(10,8,6,10,5)
#s11 <- c(1,1,2,2,2)
#x2 <- c(8,7,6,10,12,11,6,8)
#s22 <- c(1,1,1,1,2,2,2,2)
#y1 <- c(10,5,7,9,12,6)
#s1 <- c(1,1,1,2,2,2)
#y2 <- c(8,9,10,7,8,11)
#s2 <- c(1,1,1,2,2,2)
#analyze_raw_data(x1, x2, y1, y2, s11=s11, s22=s22, s1=s1, s2=s2,
#                 xi=c(0.5,0.5), nstrata=2)
}
