% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct2stage.r
\docType{data}
\name{imap}
\alias{imap}
\title{A data set available with ct2stage}
\description{
Here's something about the trial that this data set was taken 
from.
}
\references{
Moore et al. (2013) Genetics 195:1077-1086
(\href{http://www.ncbi.nlm.nih.gov/pubmed/23979570}{PubMed})
}
\keyword{data}

