% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct2stage.r
\docType{data}
\name{imap_strat}
\alias{imap_strat}
\title{Stratified summary data for the IMAP study}
\description{
Data set is from the Improving Management of Abnormal Pap Smears study,
which used a doubly randomized preference trial design to evaluate psychosocial 
outcomes in women found to have atypical cells in a Pap Smear. Two systems for
managing the atypical cells were tested (repeated Pap smears or HCV triage) and a 
doubly randomized design was used to evaluate the role of patient preference. In
this data set, patients are stratified by STAI score. Stratum 1 includes patients 
with low STAI scores (<=10 on STAI averaged across whole study); stratum 2 includes
patients wiht high STAI scores (>10). The data set provides mean, standard 
deviation and sample sizes of the SF36 outcome for each treatment in both the 
choice and random arms for both strata.
}
\references{
McCaffery et al. (2010) BMJ 340:b4491.
(\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2827716/}{PubMed})
}
\keyword{data}

