% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-size.r
\name{n_sel}
\alias{n_sel}
\title{Selection Effect Sample Size}
\usage{
n_sel(power, phi, sigma2, delta_pi, delta_nu, alpha = 0.05, theta = 0.5,
  xi = 1, nstrata = 1)
}
\arguments{
\item{power}{desired study power. Should be numeric value between 0 and 1.}

\item{phi}{the proportion of patients preferring treatment 1. Should be
numeric value between 0 and 1. If study is stratified, should be
vector with length equal to the number of strata in the study.}

\item{sigma2}{variance estimate. Should be positive numberic values. If study
is stratified, should be vector of within-stratum variances 
with length equal to the number of strata in the study.}

\item{delta_pi}{overall study preference effect.}

\item{delta_nu}{overall study selection effect.}

\item{alpha}{desired type I error rate.}

\item{theta}{proportion of patients assigned to choice arm in the initial
randomization. Should be numeric value between
0 and 1 (default=0.5).}

\item{xi}{a numeric vector of the proportion of patients in each stratum. 
Length of vector should equal the number of strata in the study and 
sum of vector should be 1. All vector elements should be numeric 
values between 0 and 1. Default is 1 (i.e. unstratified design).}

\item{nstrata}{number of strata. Default is 1 (i.e. unstratified design).}
}
\description{
Calculates the sample size required to detect a given selection effect 
in a two-stage randomized clinical trial
}
\examples{
# Unstratified
n_sel(power=0.8, phi=0.6, sigma2=1, delta_pi=1, delta_nu=0.5)
# Stratified
n_sel(power=0.8, phi=c(0.5, 0.5), sigma2=c(1, 1), delta_pi=1, 
 delta_nu=0.5,xi=c(0.3,0.7),nstrata=2)
}
\references{
Turner et al. (2014) Medical Decision Making 34:711-719.
}

