% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-size.r
\name{strat_analysis_raw}
\alias{strat_analysis_raw}
\title{Stratified Analysis Function: Raw Data}
\usage{
strat_analysis_raw(x1, s11, x2, s22, y1, s1, y2, s2, xi = c(0.5, 0.5),
  nstrata = 2)
}
\arguments{
\item{x1}{vector of responses for patients choosing treatment 1}

\item{s11}{vector of stratum membership for patients choosing treatment 1. 
Should be a vector of the same length as x1 with the number of
unique values equal to the number of strata.}

\item{x2}{vector of responses for patients choosing treatment 2}

\item{s22}{vector of stratum membership for patients choosing treatment 2. 
Should be a vector of the same length as x2 with the number of
unique values equal to the number of strata.}

\item{y1}{vector of responses for patients randomized to treatment 1}

\item{s1}{vector of stratum membership for patients randomized to treatment 1. 
Should be a vector of the same length as y1 with the number of
unique values equal to the number of strata.}

\item{y2}{vector of responses for patients randomized to treatment 2.}

\item{s2}{vector of stratum membership for patients randomized to treatment 2. 
Should be a vector of the same length as y2 with the number of
unique values equal to the number of strata.}

\item{xi}{a numeric vector of the proportion of patients in each stratum. 
Length of vector should equal the number of strata in the study and 
sum of vector should be 1. All vector elements should be numeric 
values between 0 and 1.}

\item{nstrata}{number of strata (default=2)}
}
\description{
Computes the test statistic and p-value for the preference, selection, and 
treatment effects for the stratified two-stage randomized trial using
provided raw data
}
\examples{
x1<-c(10,8,6,10,5)
s11<-c(1,1,2,2,2)
x2<-c(8,7,6,10,12,11,6,8)
s22<-c(1,1,1,1,2,2,2,2)
y1<-c(10,5,7,9,12,6)
s1<-c(1,1,1,2,2,2)
y2<-c(8,9,10,7,8,11)
s2<-c(1,1,1,2,2,2)
}

