% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-size.r
\name{strat_analysis_summary}
\alias{strat_analysis_summary}
\title{Stratified Analysis Function: Summary Data}
\usage{
strat_analysis_summary(x1mean, x1var, m1, x2mean, x2var, m2, y1mean, y1var, n1,
  y2mean, y2var, n2, xi = c(0.5, 0.5), nstrata = 2)
}
\arguments{
\item{x1mean}{vector of mean of responses for patients choosing treatment 1.
Length of vector should be equal to number of strata.}

\item{x1var}{vector of variance of responses for patients choosing 
treatment 1. Length of vector should be equal to number of 
strata.}

\item{m1}{vector of number of patients choosing treatment 1. Length of 
vector should be equal to number of strata.}

\item{x2mean}{vector of mean of responses for patients choosing treatment 2.
Length of vector should be equal to number of strata.}

\item{x2var}{vector of variance of responses for patients choosing 
treatment 2. Length of vector should be equal to number of
strata.}

\item{m2}{vector of number of patients choosing treatment 2. Length of 
vector should be equal to number of strata.}

\item{y1mean}{vector of mean of responses for patients randomized to 
treatment 1. Length of vector should be equal to number of 
strata.}

\item{y1var}{vector of variance of responses for patients randomized to 
treatment 1. Length of vector should be equal to number of 
strata.}

\item{n1}{vector of number of patients randomized to treatment 1. Length 
of vector should be equal to number of strata.}

\item{y2mean}{vector of mean of responses for patients randomized to 
treatment 2. Length of vector should be equal to number of 
strata.}

\item{y2var}{vector of variance of responses for patients randomized to 
treatment 2. Length of vector should be equal to number of 
strata.}

\item{n2}{vector of number of patients randomized to treatment 2. Length 
of vector should be equal to number of strata.}

\item{xi}{a numeric vector of the proportion of patients in each stratum. 
Length of vector should equal the number of strata in the study and 
sum of vector should be 1. All vector elements should be numeric 
values between 0 and 1.}

\item{nstrata}{number of strata (default=2)}
}
\description{
Computes the test statistic and p-value for the preference, selection, and 
treatment effects for the stratified two-stage randomized trial using
provided summary data
}
\examples{
x1<-c(10,8,6,10,5)
s11<-c(1,1,2,2,2)
x2<-c(8,7,6,10,12,11,6,8)
s22<-c(1,1,1,1,2,2,2,2)
y1<-c(10,5,7,9,12,6)
s1<-c(1,1,1,2,2,2)
y2<-c(8,9,10,7,8,11)
s2<-c(1,1,1,2,2,2)
}

