% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-size.r
\name{strat_overall}
\alias{strat_overall}
\title{Stratified Overall Sample Size}
\usage{
strat_overall(beta, phi, sigma2, delta_pi, delta_nu, delta_tau, alpha = 0.05,
  theta = 0.5, xi = c(0.5, 0.5), nstrata = 2)
}
\arguments{
\item{beta}{desired study power. Should be numeric value between 0 and 1.}

\item{phi}{vector of the proportion of patients preferring treatment 1 within
each stratum. Length of vector should equal number of strata 
in the study. Vector elements should be numeric values between 0
and 1.}

\item{sigma2}{vector of within-stratum variances. Length of vector should 
equal number of strata in the study. Vector elements should be
positive numeric values.}

\item{delta_pi}{overall study preference effect.}

\item{delta_nu}{overall study selection effect.}

\item{delta_tau}{overall study treatment effect.}

\item{alpha}{desired type I error rate.}

\item{theta}{proportion of patients assigned to choice arm in the initial
randomization. Should be numeric value between
0 and 1 (default=0.5).}

\item{xi}{a numeric vector of the proportion of patients in each stratum. 
Length of vector should equal the number of strata in the study and 
sum of vector should be 1. All vector elements should be numeric 
values between 0 and 1.}

\item{nstrata}{number of strata (default=2).}
}
\description{
Calculates the sample size required to detect a given set of effects 
in a stratified two-stage randomized clinical trial
}
\examples{
# Put example code here.
strat_overall(beta=0.8, phi=c(0.5, 0.5), sigma2=c(1, 1), delta_pi=1, 
 delta_nu=0.5, delta_tau=0.5)
}

