% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-size.r
\name{strat_preference}
\alias{strat_preference}
\title{Stratified Preference Effect Sample Size}
\usage{
strat_preference(zbeta, phi, sigma2, delta_pi, delta_nu,
  zalpha = qnorm(0.975), theta = 0.5, xi = c(0.5, 0.5), nstrata = 2)
}
\arguments{
\item{zbeta}{power Z value.}

\item{phi}{vector of preference rate of treatment 1 in each stratum.}

\item{sigma2}{vector of within-stratum variances.}

\item{delta_pi}{preference effect.}

\item{delta_nu}{selection effect.}

\item{zalpha}{type I error Z value.}

\item{theta}{proportion of patients assigned to choice arm (default=0.5).}

\item{xi}{a numeric vector of proportion of patients in each stratum.}

\item{nstrata}{number of strata (default=2).}
}
\description{
Calculates the sample size required to detect a given preference effect 
in a stratified two-stage randomized clinical trial
}
\examples{
strat_preference(1.282, c(0.5, 0.5), c(1, 1), c(1, 1), c(0, 0))
}

