% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-size.r
\name{strat_sel_pwr}
\alias{strat_sel_pwr}
\title{Stratified Selection Effect Power Calculation}
\usage{
strat_sel_pwr(N, phi, sigma2, delta_pi, delta_nu, alpha = 0.05, theta = 0.5,
  xi = c(0.5, 0.5), nstrata = 2)
}
\arguments{
\item{N}{overall study sample size.}

\item{phi}{vector of the proportion of patients preferring treatment 1 within
each stratum. Length of vector should equal number of strata 
in the study. Vector elements should be numeric values between 0
and 1.}

\item{sigma2}{vector of within-stratum variances. Length of vector should 
equal number of strata in the study. Vector elements should be
positive numeric values.}

\item{delta_pi}{overall study preference effect.}

\item{delta_nu}{overall study selection effect.}

\item{alpha}{desired type I error rate.}

\item{theta}{proportion of patients assigned to choice arm in the initial
randomization. Should be numeric value between
0 and 1 (default=0.5).}

\item{xi}{a numeric vector of the proportion of patients in each stratum. 
Length of vector should equal the number of strata in the study and 
sum of vector should be 1. All vector elements should be numeric 
values between 0 and 1.}

\item{nstrata}{number of strata (default=2).}
}
\description{
Calculates the study power to detect the selection effect given a particular 
sample size in an stratified two-stage randomized clinical trial
}
\examples{
strat_sel_pwr(N=300, phi=c(0.5,0.6), sigma2=c(1,1), delta_pi=0.5, 
               delta_nu=0.5)
}

