% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-size.r
\name{strat_selection}
\alias{strat_selection}
\title{Stratified Selection Effect}
\usage{
strat_selection(zalpha = qnorm(0.975), zbeta, theta = 0.5, alpha, sigma2,
  delta_pi, delta_nu, phi = c(0.5, 0.5), nstrata = 2)
}
\arguments{
\item{zalpha}{Fill this in.}

\item{zbeta}{Fill this in.}

\item{theta}{Fill this in.}

\item{alpha}{Fill this in.}

\item{sigma2}{Fill this in.}

\item{delta_pi}{Fill this in.}

\item{delta_nu}{Fill this in.}

\item{phi}{Fill this in.}

\item{nstrata}{Fill this in.}
}
\description{
Put a longer description of what the function does here.
}
\examples{
# Put example code here.
rnorm(10)
}

