% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-size.r
\name{strat_treatment}
\alias{strat_treatment}
\title{Stratified Treatment Effect Sample Size}
\usage{
strat_treatment(zbeta, phi, sigma2, delta_tau, zalpha = qnorm(0.975),
  theta = 0.5, xi = c(0.5, 0.5), nstrata = 2)
}
\arguments{
\item{zbeta}{power Z value.}

\item{phi}{vector of preference rate of treatment 1 in each stratum.}

\item{sigma2}{vector of within-stratum variances.}

\item{delta_tau}{treatment effect.}

\item{zalpha}{type I error Z value.}

\item{theta}{proportion of patients assigned to choice arm (default=0.5).}

\item{xi}{a numeric vector of proportion of patients in each stratum.}

\item{nstrata}{number of strata (default=2).}
}
\description{
Calculates the sample size required to detect a given treatment effect 
in a stratified two-stage randomized clinical trial
}
\examples{
# Put example code here.
rnorm(10)
}

