% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-size.r
\name{strat_trt_effect}
\alias{strat_trt_effect}
\title{Stratified Treatment Effect Back Calculation}
\usage{
strat_trt_effect(N, power, sigma2, alpha = 0.05, theta = 0.5, xi = c(0.5,
  0.5), nstrata = 2)
}
\arguments{
\item{N}{overall study sample size.}

\item{power}{desired study power. Should be numeric value between 0 and 1.}

\item{sigma2}{variance estimate. Should be a positive numeric value.}

\item{alpha}{desired type I error rate.}

\item{theta}{proportion of patients assigned to choice arm in the initial
randomization. Should be numeric value between
0 and 1 (default=0.5).}

\item{xi}{a numeric vector of the proportion of patients in each stratum. 
Length of vector should equal the number of strata in the study and 
sum of vector should be 1. All vector elements should be numeric 
values between 0 and 1.}

\item{nstrata}{number of strata (default=2).}
}
\description{
Calculates the treatment effect that can be detected given a desired study 
power and overall study sample size for the stratified two-stage 
randomized design
}
\examples{
strat_trt_effect(N=300,power=0.9,sigma2=c(1,0.8))
}

