% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-size.r
\name{strat_trt_pwr}
\alias{strat_trt_pwr}
\title{Stratified Treatment Effect Power Calculation}
\usage{
strat_trt_pwr(N, sigma2, delta_tau, alpha = 0.05, theta = 0.5, xi = c(0.5,
  0.5), nstrata = 2)
}
\arguments{
\item{N}{overall study sample size.}

\item{sigma2}{vector of within-stratum variances. Length of vector should 
equal number of strata in the study. Vector elements should be
positive numeric values.}

\item{delta_tau}{overall study treatment effect.}

\item{alpha}{desired type I error rate..}

\item{theta}{proportion of patients assigned to choice arm in the initial
randomization. Should be numeric value between
0 and 1 (default=0.5).}

\item{xi}{a numeric vector of the proportion of patients in each stratum. 
Length of vector should equal the number of strata in the study and 
sum of vector should be 1. All vector elements should be numeric 
values between 0 and 1.}

\item{nstrata}{number of strata (default=2).}
}
\description{
Calculates the study power to detect the treatment effect given a particular 
sample size in an stratified two-stage randomized clinical trial
}
\examples{
strat_trt_pwr(N=300, sigma2=c(1,1), delta_tau=0.5)
}

