% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-size.r
\name{unstrat_analysis}
\alias{unstrat_analysis}
\title{Unstratified Analysis Function}
\usage{
unstrat_analysis(x1, x2, y1, y2)
}
\arguments{
\item{x1}{vector of responses for patients choosing treatment 1}

\item{x2}{vector of responses for patients choosing treatment 2}

\item{y1}{vector of responses for patients randomized to treatment 1}

\item{y2}{vector of responses for patients randomized to treatment 2}
}
\description{
Computes the test statistic and p-value for the preference, selection, and 
treatment effects for the unstratified two-stage randomized trial
}
\examples{
x1<-c(10,9,5,12,14)
x2<-c(16,12,14,10,8,9,11)
y1<-c(10,6,5,14,8)
y2<-c(12,10,15,11,11)
unstat_analysis(x1,x2,y1,y2)
}

