% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-size.r
\name{unstrat_analysis_summary}
\alias{unstrat_analysis_summary}
\title{Unstratified Analysis Function: Summary Data}
\usage{
unstrat_analysis_summary(x1mean, x1var, m1, x2mean, x2var, m2, y1mean, y1var,
  n1, y2mean, y2var, n2)
}
\arguments{
\item{x1mean}{mean of responses for patients choosing treatment 1}

\item{x1var}{variance of responses for patients choosing treatment 1}

\item{m1}{number of patients choosing treatment 1}

\item{x2mean}{mean of responses for patients choosing treatment 2}

\item{x2var}{variance of responses for patients choosing treatment 2}

\item{m2}{number of patients choosing treatment 2}

\item{y1mean}{mean of responses for patients randomized to treatment 1}

\item{y1var}{variance of responses for patients randomized to treatment 1}

\item{n1}{number of patients randomized to treatment 1}

\item{y2mean}{mean of responses for patients randomized to treatment 2}

\item{y2var}{variance of responses for patients randomized to treatment 2}

\item{n2}{number of patients randomized to treatment 2}
}
\description{
Computes the test statistic and p-value for the preference, selection, and 
treatment effects for the unstratified two-stage randomized trial from 
provided summary data
}
\examples{
x1<-c(10,9,5,12,14)
x2<-c(16,12,14,10,8,9,11)
y1<-c(10,6,5,14,8,10)
y2<-c(12,10,15,11,11,13)
unstrat_analysis(x1,x2,y1,y2)
}

