% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-size.r
\name{unstrat_pref_pwr}
\alias{unstrat_pref_pwr}
\title{Unstratified Preference Effect Power Calculation}
\usage{
unstrat_pref_pwr(N, phi, sigma2, delta_pi, delta_nu, alpha = 0.05,
  theta = 0.5)
}
\arguments{
\item{N}{overall study sample size.}

\item{phi}{proportion of patients preferring treatment 1. Should be numeric
value between 0 and 1.}

\item{sigma2}{variance estimate. Should be a positive numeric value.}

\item{delta_pi}{overall study preference effect.}

\item{delta_nu}{overall study selection effect.}

\item{alpha}{desired type I error rate.}

\item{theta}{proportion of patients assigned to choice arm in the initial
randomization. Should be numeric value between
0 and 1 (default=0.5).}
}
\description{
Calculates the study power to detect the preference effect given a particular 
sample size in an unstratified two-stage randomized clinical trial
}
\examples{
unstrat_pref_pwr(N=300, phi=0.5, sigma2=1, delta_pi=0.5, delta_nu=0.5)
}

