% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-size.r
\name{unstrat_preference}
\alias{unstrat_preference}
\title{Unstratified Preference Effect}
\usage{
unstrat_preference(zalpha = qnorm(0.975), zbeta, theta = 0.5, alpha,
  sigma2 = 1, delta_pi, delta_nu)
}
\arguments{
\item{zalpha}{Fill this in.}

\item{zbeta}{Fill this in.}

\item{theta}{Fill this in.}

\item{alpha}{Fill this in.}

\item{sigma2}{Fill this in.}

\item{delta_pi}{Fill this in.}

\item{delta_nu}{Fill this in.}
}
\description{
Put a longer description of what the function does here.
}
\examples{
# Put example code here.
rnorm(10)
}

