% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-size.r
\name{unstrat_selection}
\alias{unstrat_selection}
\title{Unstratified Selection Effect Sample Size}
\usage{
unstrat_selection(zbeta, phi, sigma2, delta_pi, delta_nu,
  zalpha = qnorm(0.975), theta = 0.5)
}
\arguments{
\item{zbeta}{power Z value.}

\item{phi}{preference rate of treatment 1.}

\item{sigma2}{variance.}

\item{delta_pi}{preference effect.}

\item{delta_nu}{selection effect.}

\item{zalpha}{type I error Z value.}

\item{theta}{proportion of patients assigned to choice arm (default=0.5).}
}
\description{
Calculates the sample size required to detect a given selection effect 
in an unstratified two-stage randomized clinical trial
}
\examples{
# Put example code here.
rnorm(10)
}

