% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-size.r
\name{unstrat_treatment}
\alias{unstrat_treatment}
\title{Unstratified Treatment Effect}
\usage{
unstrat_treatment(zalpha = qnorm(0.0975), zbeta, theta = 0.5, sigma2 = 1,
  delta_tau)
}
\arguments{
\item{zalpha}{Fill this in.}

\item{zbeta}{Fill this in.}

\item{theta}{Fill this in.}

\item{sigma2}{Fill this in.}

\item{delta_tau}{Fill this in.}

\item{alpha}{Fill this in.}
}
\description{
Put a longer description of what the function does here.
}
\examples{
# Put example code here.
rnorm(10)
}

