% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-size.r
\name{unstrat_trt_effect}
\alias{unstrat_trt_effect}
\title{Unstratified Treatment Effect Back Calculation}
\usage{
unstrat_trt_effect(N, power, sigma2, alpha = 0.05, theta = 0.5)
}
\arguments{
\item{N}{overall study sample size.}

\item{power}{desired study power. Should be numeric value between 0 and 1.}

\item{sigma2}{variance estimate. Should be a positive numeric value.}

\item{alpha}{desired type I error rate.}

\item{theta}{proportion of patients assigned to choice arm in the initial
randomization. Should be numeric value between
0 and 1 (default=0.5).}
}
\description{
Calculates the treatment effect that can be detected given a desired study 
power and overall study sample size for the unstratified two-stage 
randomized design
}
\examples{
unstrat_trt_effect(N=300,power=0.9,sigma2=1)
}

