% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-size.r
\name{unstrat_trt_pwr}
\alias{unstrat_trt_pwr}
\title{Unstratified Treatment Effect Power Calculation}
\usage{
unstrat_trt_pwr(N, sigma2, delta_tau, alpha = 0.05, theta = 0.5)
}
\arguments{
\item{N}{overall study sample size.}

\item{sigma2}{variance estimate. Should be a positive numeric value.}

\item{delta_tau}{overall study treatment effect.}

\item{alpha}{desired type I error rate..}

\item{theta}{proportion of patients assigned to choice arm in the initial
randomization. Should be numeric value between
0 and 1 (default=0.5).}
}
\description{
Calculates the study power to detect the treatment effect given a particular 
sample size in an unstratified two-stage randomized clinical trial
}
\examples{
unstrat_trt_pwr(N=300, sigma2=1, delta_tau=0.5)
}

