% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-size.r
\name{unstrat_treatment}
\alias{unstrat_treatment}
\title{Unstratified Treatment Effect Sample Size}
\usage{
unstrat_treatment(beta, sigma2, delta_tau, alpha = 0.05, theta = 0.5)
}
\arguments{
\item{beta}{desired study power. Should be numeric value between 0 and 1.}

\item{sigma2}{variance estimate. Should be a positive numeric value.}

\item{delta_tau}{overall study treatment effect.}

\item{alpha}{desired type I error rate.}

\item{theta}{proportion of patients assigned to choice arm in the initial
randomization. Should be numeric value between
0 and 1 (default=0.5).}
}
\description{
Calculates the sample size required to detect a given treatment effect 
in an unstratified two-stage randomized clinical trial
}
\examples{
# Put example code here.
unstrat_treatment(beta=0.8, sigma2=1, delta_tau=0.5)
}

