% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-size.r
\name{unstrat_treatment}
\alias{unstrat_treatment}
\title{Unstratified Treatment Effect Sample Size}
\usage{
unstrat_treatment(zbeta, sigma2 = 1, delta_tau, zalpha = qnorm(0.0975),
  theta = 0.5)
}
\arguments{
\item{zbeta}{power Z value.}

\item{sigma2}{variance.}

\item{delta_tau}{treatment effect.}

\item{zalpha}{type I error Z value.}

\item{theta}{proportion of patients assigned to choice arm (default=0.5).}

\item{phi}{preference rate of treatment 1.}
}
\description{
Calculates the sample size required to detect a given treatment effect 
in an unstratified two-stage randomized clinical trial
}
\examples{
# Put example code here.
rnorm(10)
}

